+++ docs/pages/components/loading/Loading.vue
@@ -6,7 +6,10 @@
 
         <Example :component="ExSimple" :code="ExSimpleCode">
             <p>The Loading will be closed after about 10 seconds, by pressing escape or by clicking outside.</p>
-            <p>Use <code>:is-full-page="false"</code> to limit loader to its container.</p>
+            <p>
+                Use <code>:is-full-page="false"</code> to limit loader to its container.<br>
+                In this case, the container element should be positioned as <code>position: relative</code>.
+            </p>
         </Example>
 
         <Example :component="ExProgrammatically" :code="ExProgrammaticallyCode" title="Programmatically opening">