+++ src/components/input/Input.vue
@@ -45,7 +45,7 @@
             @click.native="togglePasswordVisibility"/>
 
         <small
-            v-if="maxlength && hasCounter"
+            v-if="maxlength && hasCounter && type !== 'number'"
             class="help counter"
             :class="{ 'is-invisible': !isFocused }">
             {{ valueLength }} / {{ maxlength }}
@@ -153,7 +153,12 @@
              * Get value length
              */
             valueLength() {
-                return this.newValue ? this.newValue.length : 0
+                if (typeof this.newValue === 'string') {
+                    return this.newValue.length
+                } else if (typeof this.newValue === 'number') {
+                    return this.newValue.toString().length
+                }
+                return 0
             }
         },
         watch: {