+++ src/components/radio/Radio.vue
@@ -12,7 +12,7 @@
             :disabled="disabled"
             :name="name"
             :value="nativeValue">
-        <span class="check"/>
+        <span class="check" :class="type" />
         <span class="control-label"><slot/></span>
     </label>
 </template>
@@ -23,6 +23,7 @@
         props: {
             value: [String, Number, Boolean, Function, Object, Array, Symbol],
             nativeValue: [String, Number, Boolean, Function, Object, Array, Symbol],
+            type: String,
             disabled: Boolean,
             name: String,
             size: String